Fix gw-refresh-worker-type-definitions-from-occ.sh

+++ gw-refresh-worker-type-definitions-from-occ.sh
@@ -1,15 +1,7 @@
 #!/bin/bash -xve
 
-GPG_NO_TTY="$(mktemp -t gpg-no-tty.XXXXXXXXXX)"
-{
-  echo '#!/bin/bash'
-  echo 'gpg --no-tty "${@}"'
-  echo 'exit $?'
-} >> "${GPG_NO_TTY}"
-chmod u+x "${GPG_NO_TTY}"
-
-function git_no_tty {
-    git -c "gpg.program=${GPG_NO_TTY}" "${@}"
+function git_no_gpg {
+    git -c "commit.gpgsign=false" "${@}"
 }
 
 # export PATH="/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin:/usr/local/go/bin"
@@ -17,11 +9,18 @@ source ~/gw-refresh-worker-type-definitions-from-occ.env
 TMP_GW_CHECKOUT="$(mktemp -d -t refresh-gw-worker-types.XXXXXXXXXX)"
 cd "${TMP_GW_CHECKOUT}"
 git clone git@github.com:taskcluster/generic-worker
-generic-worker/worker_types/generate_occ_userdata.sh
 cd generic-worker/worker_types
-git_no_tty add .
-git_no_tty commit -m "Refreshed gecko worker type definitions from latest OCC manifests"
-git_no_tty push origin master
+./generate_occ_userdata.sh
+git_no_gpg add .
+if ! git_no_gpg commit -m "Refreshed gecko worker type definitions from latest OCC manifests"; then
+  say -v Daniel "No OCC changes"
+  exit 64
+fi
+if ! git_no_gpg push origin master; then
+  say -v Daniel "Problem updating OCC manifests. Pete, what is going on?"
+  exit 65
+fi
 say -v Daniel "Updated OCC manifests"
 cd ~
 rm -rf "${TMP_GW_CHECKOUT}"
+exit 0
